+++ test/test_transform_id.py
@@ -1,5 +1,5 @@
 import unittest
-from numpy import array
+from numpy import array, int8
 from numpy.testing import assert_array_equal
 import pypuf_helper as ph
 
@@ -11,7 +11,7 @@ class TestTransformID(unittest.TestCase):
         input_test = array([
             [1, -1,  1, 1],
             [-1, -1,  1, 1]
-            ])
+            ], dtype=int8)
         result = array([
             [[1, -1,  1, 1],
              [1, -1,  1, 1],
@@ -19,7 +19,7 @@ class TestTransformID(unittest.TestCase):
             [[-1, -1,  1, 1],
              [-1, -1,  1, 1],
              [-1, -1,  1, 1]]
-            ])
+            ], dtype=int8)
         assert_array_equal(
             ph.transform_id(input_test, k),
             result,
